Refactor only functions
